# App/openapi.yaml
openapi: 3.0.3
info:
  title: Voenkom API
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
...
paths:
  /api/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful, returns user and JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
  /api/me:
    get:
      security:
        - BearerAuth: []
      summary: Get current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
  /api/users:
    get:
      security:
        - BearerAuth: []
      summary: List users
      responses:
        '200':
          description: Users array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      security:
        - BearerAuth: []
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: Created user id
  /api/users/{id}:
    delete:
      security:
        - BearerAuth: []
      summary: Delete user (non-commissar)
...
  /api/appointments:
    get:
      security:
        - BearerAuth: []
      summary: List appointments
    post:
      security:
        - BearerAuth: []
      summary: Create appointment
...
  /api/calendar:
    get:
      security:
        - BearerAuth: []
      summary: List calendar events
    post:
      security:
        - BearerAuth: []
      summary: Create calendar event
...